make table for ticket price
I got rid of the extra page and made it to where a table with ticket prices renders on the main page when a user enters a resort. The input box isn't currently working with autocomplete because for some reason it wont load the table and do autocomplete at the same time.